Refactoring: extract class, create "GrafanaBackend" class
